refactor(ui): extract InfoTooltip as shared component (keyboard-accessible)
SignalSubsection ⓘ 버튼이 title + sr-only만으로 구성되어 키보드 사용자가
툴팁 콘텐츠를 볼 수 없었음. AnalysisPanel의 완전한 InfoTooltip
(click/pointer/portal/focus-visible ring)을 src/components/ui/InfoTooltip.tsx로
분리해 양쪽에서 재사용.

- NEW: src/components/ui/InfoTooltip.tsx (focus-visible:ring 기본 포함)
- AnalysisPanel: 로컬 InfoTooltip 제거 + shared import
- SignalSubsection: 키보드 접근 가능한 shared InfoTooltip 채택

MISTAKES.md Accessibility #4 준수. Reviewer: claude[bot] 1 blocker, 1 suggestion.
