+++ src/components/analysis/AnalysisPanel.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { useEffect, useId, useRef, useState } from 'react';
-import { createPortal } from 'react-dom';
+import { useEffect, useRef, useState } from 'react';
+import { InfoTooltip } from '@/components/ui/InfoTooltip';
 import type React from 'react';
 import type {
     ActionRecommendation,
@@ -37,7 +37,6 @@ import { resolveStrengthDisplay } from '@/components/analysis/utils/signalUtils'
 import { AnalysisProgress } from '@/components/analysis/AnalysisProgress';
 import { AnalysisToast } from '@/components/analysis/AnalysisToast';
 import { AdBanner } from '@/components/analysis/AdBanner';
-import { useOnClickOutside } from '@/components/hooks/useOnClickOutside';
 import type { CooldownNotice } from '@/components/symbol-page/hooks/useAnalysis';
 import { TRENDLINE_DIRECTION_LABEL } from '@/components/trendline/constants';
 
@@ -428,120 +427,6 @@ function ConfidenceBadge({ confidenceWeight }: ConfidenceBadgeProps) {
     );
 }
 
-const TOOLTIP_VIEWPORT_PADDING = 8;
-
-interface TooltipPosition {
-    top: number;
-    left: number;
-}
-
-const TOOLTIP_GAP = 6;
-
-function getTooltipPosition(
-    triggerRect: DOMRect,
-    tooltipEl: HTMLElement
-): TooltipPosition {
-    const tooltipRect = tooltipEl.getBoundingClientRect();
-    const aboveTop = triggerRect.top - tooltipRect.height - TOOLTIP_GAP;
-    const top =
-        aboveTop < TOOLTIP_VIEWPORT_PADDING
-            ? triggerRect.bottom + TOOLTIP_GAP
-            : aboveTop;
-    const rawLeft =
-        triggerRect.left + triggerRect.width / 2 - tooltipRect.width / 2;
-    const maxLeft =
-        window.innerWidth - tooltipRect.width - TOOLTIP_VIEWPORT_PADDING;
-    const left = Math.max(TOOLTIP_VIEWPORT_PADDING, Math.min(rawLeft, maxLeft));
-
-    return { top, left };
-}
-
-interface InfoTooltipProps {
-    children: React.ReactNode;
-}
-
-function InfoTooltip({ children }: InfoTooltipProps) {
-    const tooltipId = useId();
-    const [open, setOpen] = useState(false);
-    const [positioned, setPositioned] = useState(false);
-    const triggerRef = useRef<HTMLButtonElement>(null);
-    const tooltipRef = useRef<HTMLDivElement>(null);
-    const [position, setPosition] = useState<TooltipPosition>({
-        top: 0,
-        left: 0,
-    });
-
-    useOnClickOutside([triggerRef, tooltipRef], () => setOpen(false), {
-        enabled: open,
-    });
-
-    const handleClick = (): void => {
-        if (open) {
-            setOpen(false);
-            setPositioned(false);
-        } else {
-            setOpen(true);
-        }
-    };
-
-    const handlePointerEnter = (e: React.PointerEvent): void => {
-        if (e.pointerType === 'touch') return;
-        setOpen(true);
-    };
-
-    const handlePointerLeave = (e: React.PointerEvent): void => {
-        if (e.pointerType === 'touch') return;
-        setOpen(false);
-        setPositioned(false);
-    };
-
-    return (
-        <>
-            <button
-                ref={triggerRef}
-                type="button"
-                aria-describedby={open ? tooltipId : undefined}
-                onClick={handleClick}
-                onPointerEnter={handlePointerEnter}
-                onPointerLeave={handlePointerLeave}
-                className="text-secondary-600 hover:text-secondary-400 ml-1 cursor-help text-xs leading-none transition-colors"
-            >
-                ⓘ
-            </button>
-            {open &&
-                createPortal(
-                    <div
-                        ref={el => {
-                            tooltipRef.current = el;
-                            if (el && triggerRef.current) {
-                                const triggerRect =
-                                    triggerRef.current.getBoundingClientRect();
-                                const pos = getTooltipPosition(triggerRect, el);
-                                if (
-                                    pos.top !== position.top ||
-                                    pos.left !== position.left
-                                ) {
-                                    setPosition(pos);
-                                }
-                                if (!positioned) setPositioned(true);
-                            }
-                        }}
-                        id={tooltipId}
-                        role="tooltip"
-                        className="bg-secondary-800 border-secondary-600 fixed z-9999 rounded border p-2 text-xs leading-relaxed shadow-lg"
-                        style={{
-                            top: position.top,
-                            left: position.left,
-                            visibility: positioned ? 'visible' : 'hidden',
-                        }}
-                    >
-                        {children}
-                    </div>,
-                    document.body
-                )}
-        </>
-    );
-}
 
 interface ConfluenceInfoProps {
     level: ClusteredKeyLevel;
